+++ src/main/java/com/example/kahoot/data/executors/DbAddQuestion.kt
@@ -3,8 +3,8 @@ package com.example.kahoot.data.executors
 import com.example.kahoot.data.DatabaseInterface
 import com.example.kahoot.domain.model.KahootQuestion
 
-class DbAddQuestion(private val db: DatabaseInterface.Executor) : DbExecutor.Base<KahootQuestion, Int> {
-    override fun execute(objects: KahootQuestion): Int {
+class DbAddQuestion(private val db: DatabaseInterface.Executor) : DbExecutor.Base<KahootQuestion, Long> {
+    override fun execute(objects: KahootQuestion): Long {
         val sql = "INSERT or IGNORE INTO Question (title,correct) VALUES (? , ?)"
         val qId = db.executeUpdate(sql, objects.question, objects.correct)
         objects.answers.forEach {
